Size Reduction!
makefile : added SAVE_SPACE define, this define is meant to call the
SAVE_SPACE define in all code, reducing code size so it can fit on a
ATMega8 if we want (not that it can for now, because of the lack of pins
:P )

GBC_API : added the dumping of ram & rom here.

gbc_Error : moved all error defines to a header so it can be used across
the program

GB_Cart core : added a MBC type byte to the game info, replaced the
function calling with said byte, removed some header checking & setting
of the RST pin and all that since it should be done by the higher up
code ( like the API ). added support for SAVE_SPACE

Main :  support for SAVE_SPACE & remove none API commands because bleh

Serial : Support for SAVE_SPACE, this kills VA support which we were
hardly using anyway